01-26 15:47 Changed _age from a int to a double

+++ P02/full_credit/pet.cpp
@@ -7,7 +7,7 @@ std::string to_string(Type pet)
 	return q[pet];
 }
 
-Pet::Pet( std::string n , int a , Type p) : _name{n} , _age{a} , pet{p} 
+Pet::Pet( std::string n , double a , Type p) : _name{n} , _age{a} , pet{p} 
 {
 
 }
